requirements.txt
aiogram==3.18.0
httpx==0.28.1
beautifulsoup4==4.13.3
python-dotenv==1.0.1
lxml==5.3.1
